+++ api/src/main/java/com/codeforcommunity/rest/ApiRouter.java
@@ -22,7 +22,8 @@ import com.codeforcommunity.util.UpdatableBCrypt;
 import com.codeforcommunity.util.JWTUtils;
 
 import com.codeforcommunity.rest.subrouter.CommonRouter;
-
+import com.codeforcommunity.rest.subrouter.EventsRouter;
+import com.codeforcommunity.rest.subrouter.UsersAuthRouter;
 import com.fasterxml.jackson.core.JsonProcessingException;
 
 import io.vertx.core.MultiMap;
@@ -66,12 +67,8 @@ import java.util.List;
 import java.util.Optional;
 
 public class ApiRouter {
-  private HashMap<String, Integer> loginmap = new HashMap<String, Integer>();
   private final IProcessor processor;
   private Vertx v;
-  boolean flag = false;
-
-  private static final UpdatableBCrypt bcrypt = new UpdatableBCrypt(12);
 
   private final JWTUtils auth;
 
@@ -93,62 +90,12 @@ public class ApiRouter {
 
     router.mountSubRouter("/", new CommonRouter(processor, auth).initializeRouter(vertx));
 
+    router.mountSubRouter("/", new UsersAuthRouter(processor, auth).initializeRouter(vertx));
+
+    router.mountSubRouter("/", new EventsRouter(processor, auth).initializeRouter(vertx));
+
     Route homeRoute = router.route("/");
     homeRoute.handler(this::handleHome);
-
-    // allows handling of POST and DELETE routes.
-    // router.route("/*").handler(BodyHandler.create());
-
-    // Users
-
-    Route loginRoute = router.post("/login");
-    loginRoute.handler(this::handleLogin);
-
-    Route signUpRoute = router.post("/signup");
-    signUpRoute.handler(this::handleSignUp);
-
-    Route getUsersRoute = router.get("/protected/users");
-    getUsersRoute.handler(this::handleGetAllUsers);
-
-    Route getUserRoute = router.get("/protected/users/:id");
-    getUserRoute.handler(this::handleGetUser);
-
-    Route updateUserRoute = router.put("/protected/users");
-    updateUserRoute.handler(this::handleUpdateUser);
-
-    Route deleteUserRoute = router.delete("/protected/users");
-    deleteUserRoute.handler(this::handleDeleteUser);
-
-    Route logoutRoute = router.route("/protected/logout");
-    logoutRoute.handler(this::handleLogout);
-
-    // Route sessionRoute = router.route("/session");
-    // sessionRoute.handler(this::handleSession);
-
-    // Events
-
-    Route getEventsRoute = router.get().path("/events");
-    getEventsRoute.handler(this::handleGetEvents);
-
-    Route createEventRoute = router.post("/admin/events");
-    createEventRoute.handler(this::handleCreateEvent);
-
-    Route getEventRoute = router.get("/events/:id");
-    getEventRoute.handler(this::handleGetEvent);
-
-    Route updateEventRoute = router.put("/admin/events/:id");
-    updateEventRoute.handler(this::handleUpdateEvent);
-
-    Route deleteEventRoute = router.delete("/admin/events/:id");
-    deleteEventRoute.handler(this::handleDeleteEvent);
-
-    // Event Check Ins
-
-    Route attendEventRoute = router.post("/protected/eventcheckin/:code");
-    attendEventRoute.handler(this::handleAttendEvent);
-
-    Route getEventUsersRoute = router.get("/protected/eventcheckin/:id");
-    getEventUsersRoute.handler(this::handleGetEventUsers);
 
     // Applicants
 
@@ -187,481 +134,11 @@ public class ApiRouter {
     return router;
   }
 
-  /**
-   * Add a handler for getting all users.
-   */
-  private void handleGetAllUsers(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    response.putHeader("content-type", "application/json");
-    List<UserReturn> users = processor.getAllUsers();
-
-    String userJson = null;
-    try {
-      userJson = JacksonMapper.getMapper().writeValueAsString(users);
-    } catch (JsonProcessingException e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-    response.end(userJson);
-  }
-
-  /**
-   * login handler
-   */
   private void handleHome(RoutingContext ctx) {
     HttpServerResponse response = ctx.response();
     response.putHeader("Content-Type", "text/html; charset=UTF-8");
-    response.end("<h1>go to /login with query string to login</h1>");
-  }
-
-  private void handleLogin(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-
-    JsonObject body = ctx.getBodyAsJson();
-    String email = "";
-    String password = "";
-
-    try {
-      email = body.getString("email");
-      password = body.getString("password");
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-
-    if (!flag && processor.validate(email, password)) {
-      UserReturn user = processor.getUserByEmail(email).get();
-
-      String token = auth.createJWT("c4c", "auth-token", user.getId(), user.getPrivilegeLevel() > 0);
-      response.putHeader("Authorization", "Bearer " + token);
-      response.setStatusCode(200).end();
-    }
-
-    /*
-     * Need to fix, because this will cause a timeout if login failure occurs 5
-     * times in any period of time which could be with 1 minute or 2 weeks
-     */
-    else { // if user fails to input correct password
-      if (loginmap.get(email) == null) { // first failure creates entry in cache
-        loginmap.put(email, 0); // initializes to 0th failure
-      } else
-        loginmap.put(email, loginmap.get(email) + 1); // increments failure counter by 1
-      if (loginmap.get(email) > 4) { // if they have failed 5 times
-        // Block the user from logging in for 5 minutes
-        flag = true;
-        final String username1 = email;
-        new Thread() {
-          public void run() {
-            try {
-              Thread.sleep(300000);
-            } catch (InterruptedException e) {
-              e.printStackTrace();
-            }
-            flag = false;
-            loginmap.put(username1, null); // initializes to 0th failure again after the 5 minutes is over
-          }
-        }.start();
-      }
-
-      response.putHeader("content-type", "text/json").setStatusCode(400).end();
-    }
-  }
-
-  private void handleSignUp(RoutingContext ctx) {
-
-    HttpServerResponse response = ctx.response();
-
-    JsonObject body = ctx.getBodyAsJson();
-
-    String email = "";
-    String encryptedPassword = "";
-    String firstName = "";
-    String lastName = "";
-    int currentYear = -1;
-    String major = "";
-    try {
-      email = body.getString("email");
-      firstName = body.getString("firstName");
-      lastName = body.getString("lastName");
-      encryptedPassword = bcrypt.hash(body.getString("password"));
-      currentYear = body.getInteger("currentYear");
-      major = body.getString("major");
-
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-
-    Optional<UserReturn> ret = Optional.empty();
-
-    if (email != null && firstName != null && lastName != null && encryptedPassword != null && currentYear != -1
-        && major != null)
-      ret = processor.addUser(email, firstName, lastName, encryptedPassword, currentYear, major);
-
-    String json = "";
-
-    try {
-      if (ret.isPresent())
-        json = JacksonMapper.getMapper().writeValueAsString(ret.get());
-    } catch (JsonProcessingException e) {
-      e.printStackTrace();
-    }
-
-    if (!json.isEmpty()) {
-      response.setStatusCode(201).putHeader("content-type", "text/json").end(json);
-    } else {
-      response.setStatusCode(400).end();
-    }
-  }
-
-  private void handleGetUser(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-    int id = -1;
-
-    try {
-      id = Integer.parseInt(request.params().get("id"));
-    } catch (NumberFormatException e) {
-      e.printStackTrace();
-      throw new MalformedParameterException("id");
-    }
-
-    Optional<UserReturn> ret = processor.getUser(id);
-    String json = "";
-    try {
-      if (ret.isPresent())
-        json = JacksonMapper.getMapper().writeValueAsString(ret.get());
-    } catch (JsonProcessingException e) {
-      e.printStackTrace();
-    }
-
-    if (!json.isEmpty()) {
-      response.setStatusCode(200).putHeader("content-type", "text/json").end(json);
-    } else {
-      response.setStatusCode(400).end();
-    }
-  }
-
-  private void handleUpdateUser(RoutingContext ctx) {
-
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-
-    JsonObject body = ctx.getBodyAsJson();
-
-    int id = -1;
-    String email = "";
-    String password = "";
-    String encryptedPassword = "";
-
-    String firstName = "";
-    String lastName = "";
-
-    int currentYear = -1;
-    String major = "";
-    try {
-      id = auth.getUserId(request);
-      email = body.getString("email");
-      firstName = body.getString("firstName");
-      lastName = body.getString("lastName");
-      password = body.getString("password");
-
-      currentYear = body.getInteger("currentYear");
-      major = body.getString("major");
-      if (!password.isEmpty())
-        encryptedPassword = bcrypt.hash(password);
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-
-    Optional<UserReturn> ret = Optional.empty();
-    if (id != 0 && email != null && firstName != null && lastName != null && encryptedPassword != null
-        && currentYear != -1 && major != null)
-      ret = processor.updateUser(id, email, firstName, lastName, encryptedPassword, currentYear, major);
-
-    String json = "";
-    try {
-      if (ret.isPresent())
-        json = JacksonMapper.getMapper().writeValueAsString(ret.get());
-    } catch (JsonProcessingException e) {
-      e.printStackTrace();
-    }
-
-    if (!json.isEmpty()) {
-      response.setStatusCode(200).putHeader("content-type", "text/json").end(json);
-    } else {
-      response.setStatusCode(400).end();
-    }
-  }
-
-  private void handleDeleteUser(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-    int id = -1;
-    try {
-      id = auth.getUserId(request);
-      Optional<UserReturn> ret = processor.deleteUser(id);
-      String json = "";
-      try {
-        if (ret.isPresent())
-          json = JacksonMapper.getMapper().writeValueAsString(ret.get());
-      } catch (JsonProcessingException e) {
-        e.printStackTrace();
-      }
-
-      if (!json.isEmpty()) {
-        response.setStatusCode(200).putHeader("content-type", "text/json").end(json);
-      } else {
-        response.setStatusCode(400).end();
-      }
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-  }
-
-  // handling logouts requires that we save state for some JWT, if we need to keep
-  // track of state, why use stateless JWT? Wouldn't sessions be more appropriate?
-  private void handleLogout(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-    // onLogout clear expired tokens to periodically "trim" the blacklsited tokens
-    // (should probably instead daily)
-    boolean cleared = processor.clearBlacklistedTokens(auth.getDuration());
-    boolean success = processor.addBlacklistedToken(auth.getClaims(request).getId());
-
-    if (success) {
-      // ctx.reroute(ctx.request().path());
-      response.setStatusCode(201).end();
-    } else {
-      // what do we do when logout fails? (secuirty risk)
-      response.setStatusCode(500).end();
-    }
-  }
-
-  // Event Route Handlers
-  private void handleGetEvents(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    response.putHeader("content-type", "application/json");
-    List<EventReturn> events = processor.getAllEvents();
-
-    String eventsJson = null;
-    try {
-      eventsJson = JacksonMapper.getMapper().writeValueAsString(events);
-    } catch (JsonProcessingException e) {
-      e.printStackTrace();
-    }
-    response.end(eventsJson);
-  }
-
-  private void handleCreateEvent(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-    JsonObject body = ctx.getBodyAsJson();
-    String name = "";
-    String subtitle = "";
-    String description = "";
-    String imageUrl = "";
-    LocalDateTime date = null;
-    Boolean open = null;
-    String eventCode = "";
-    // for now, the input date is to the minute
-    DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-
-    try {
-      name = body.getString("name");
-      subtitle = body.getString("subtitle");
-      description = body.getString("description");
-      imageUrl = body.getString("imageUrl");
-      date = LocalDateTime.parse(body.getString("date"), formatter);
-      open = body.getBoolean("open");
-      eventCode = body.getString("code");
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-    // -----------------------DUPLICATE FOR OTHER TYPES------------------
-
-    Optional<EventReturn> result = Optional.empty();
-    if (name != null && date != null && subtitle != null && description != null && imageUrl != null && open != null
-        && eventCode != null)
-      result = processor.createEvent(name, subtitle, description, imageUrl, date, open, eventCode);
-
-    if (result.isPresent()) {
-
-      String json = "";
-      try {
-        if (result.isPresent())
-          json = JacksonMapper.getMapper().writeValueAsString(result.get());
-      } catch (JsonProcessingException e) {
-        response.setStatusCode(400).end();
-      }
-      response.setStatusCode(201).putHeader("content-type", "text/json").end(json);
-    } else {
-      response.setStatusCode(400).end();
-    }
-
-    response.setStatusCode(201);
-    // ----------------------------------END------------------------------
-
-  }
-
-  private void handleGetEvent(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-    int id = -1;
-
-    try {
-      id = Integer.parseInt(request.params().get("id"));
-    } catch (NumberFormatException e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-
-    Optional<EventReturn> ret = processor.getEvent(id);
-    String json = "";
-    try {
-      if (ret.isPresent())
-        json = JacksonMapper.getMapper().writeValueAsString(ret.get());
-    } catch (JsonProcessingException e) {
-    }
-
-    if (!json.isEmpty()) {
-      response.setStatusCode(200).putHeader("content-type", "text/json").end(json);
-    } else {
-      response.setStatusCode(400).end();
-    }
-  }
-
-  private void handleUpdateEvent(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-    JsonObject body = ctx.getBodyAsJson();
-    String name = "";
-    String subtitle = "";
-    String description = "";
-    String imageUrl = "";
-    LocalDateTime date = null;
-    Boolean open = null;
-    String eventCode = "";
-    // for now, the input date is to the minute
-    DateTimeFormatter formatter = DateTimeFormatter.ISO_LOCAL_DATE_TIME;
-    int id = -1;
-
-    try {
-      id = Integer.parseInt(request.getParam("id"));
-      name = body.getString("name");
-      subtitle = body.getString("subtitle");
-      description = body.getString("description");
-      imageUrl = body.getString("imageUrl");
-      date = LocalDateTime.parse(body.getString("date"), formatter);
-      open = body.getBoolean("open");
-      eventCode = body.getString("code");
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-
-    // -----------------------DUPLICATE FOR OTHER TYPES------------------
-
-    Optional<EventReturn> ret = Optional.empty();
-    if (name != null && date != null && subtitle != null && description != null && imageUrl != null && open != null
-        && eventCode != null)
-      ret = processor.updateEvent(id, name, subtitle, description, imageUrl, date, open, eventCode);
-
-    String json = "";
-    try {
-      if (ret.isPresent())
-        json = JacksonMapper.getMapper().writeValueAsString(ret.get());
-    } catch (JsonProcessingException e) {
-      e.printStackTrace();
-    }
-
-    if (!json.isEmpty()) {
-      response.setStatusCode(200).putHeader("content-type", "text/json").end(json);
-    } else {
-      response.setStatusCode(400).end();
-    }
-    // ----------------------------------END------------------------------
-  }
-
-  private void handleDeleteEvent(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-
-    // -----------------------DUPLICATE FOR OTHER TYPES------------------
-    int id = -1;
-    try {
-      id = Integer.parseInt(request.params().get("id"));
-      String json = "";
-
-      Optional<EventReturn> ret = processor.deleteEvent(id);
-      if (ret.isPresent())
-        json = JacksonMapper.getMapper().writeValueAsString(ret.get());
-
-      if (!json.isEmpty()) {
-        response.setStatusCode(200).putHeader("content-type", "text/json").end(json);
-      } else {
-        response.setStatusCode(400).end();
-      }
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-    // ----------------------------------END------------------------------
-  }
-
-  private void handleAttendEvent(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-
-    String eventCode = "";
-    int userId = -1;
-    boolean success = false;
-
-    try {
-      userId = auth.getUserId(request);
-      eventCode = request.params().get("code");
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-
-    if (!eventCode.isEmpty() && userId != -1)
-      success = processor.attendEvent(eventCode, userId);
-
-    if (success) {
-      response.setStatusCode(201).end();
-    } else {
-      response.setStatusCode(400).end();
-    }
-  }
-
-  private void handleGetEventUsers(RoutingContext ctx) {
-    HttpServerResponse response = ctx.response();
-    HttpServerRequest request = ctx.request();
-
-    int eventId = -1;
-    List<UserReturn> users = null;
-
-    try {
-
-      eventId = Integer.parseInt(request.params().get("id"));
-
-      users = processor.getEventUsers(eventId);
-    } catch (Exception e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-    String userJson = null;
-    try {
-      userJson = JacksonMapper.getMapper().writeValueAsString(users);
-    } catch (JsonProcessingException e) {
-      e.printStackTrace();
-      response.setStatusCode(400).end();
-    }
-    response.end(userJson);
+    response
+        .end("<a href=\"https://github.com/Code-4-Community/c4c-internal-backend/blob/master/api.md\">API Docs</a>");
   }
 
   private void handleGetApplicants(RoutingContext ctx) {
